Moved Addresses API Entity classes to Infrastructure namespace and separate files.
Decided to go with the architecture similar to that of how EF Core is configured, where the API call will have corresponding Entity classes that can be reduced to API Domain classes.
