+++ LBHFSSPublicAPI/LBHFSSPublicAPI.csproj
@@ -29,6 +29,7 @@
 
   <ItemGroup>
     <Folder Include="terraform\" />
+    <Folder Include="V1\Infrastructure\AddressesContextEntities\" />
   </ItemGroup>
 
 </Project>
