+++ LBHFSSPublicAPI/V1/Infrastructure/AddressesContextEntities/RootAPIResponseEntity.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace LBHFSSPublicAPI.V1.Infrastructure.AddressesContextEntities
+{
+    public class RootAPIResponseEntity
+    {
+        public DataEntity Data { get; set; }
+        public int StatusCode { get; set; }
+        public ErrorEntity Error { get; set; }
+    }
+}
